pyproject.toml
[tool.poetry]
name = "env-setup"
version = "0.1.0"
description = ""
authors = ["Spiros Dontas <spirosdontas@gmail.com>"]
packages = [
    {include = "setup", from = "."},
]

[tool.poetry.dependencies]
python = "^3.7"
python-dotenv = "^0.10.3"

[tool.poetry.dev-dependencies]
flake8 = "^3.7"
flake8-mypy = "^17.8"
black = {version = "^18.3-alpha.0",allows-prereleases = true}
pre-commit = "^1.17"

[tool.poetry.scripts]
debian = "setup.setup:main"

[tool.black]
line-length = 79

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
